Allow non-MLP networks in actor-critic model.
For backwards-compatability, and because I don't have anything else set up yet, the MLP is still made, but it should now be easy to write a function that applies whatever network you want on the input. There isn't a way to pass in new command-line arguments as networks kwargs, though.